add news crawler
